extract ActionKeyType
