+++ packages/flutter_onscreen_keyboard/lib/src/layouts/desktop_layout.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_onscreen_keyboard/flutter_onscreen_keyboard.dart';
+import 'package:flutter_onscreen_keyboard/src/constants/action_key_type.dart';
 
 /// A complete QWERTY desktop keyboard layout excluding the number pad.
 ///
@@ -103,25 +104,4 @@ class DesktopKeyboardLayout extends KeyboardLayout {
       ],
     ),
   ];
-}
-
-/// Defines action key type constants used in [ActionKey] widgets.
-///
-/// These strings are used to uniquely identify action buttons
-/// such as "backspace", "enter", and "shift" within the keyboard logic.
-abstract class ActionKeyType {
-  /// Identifier for the Backspace or Delete key.
-  static const backspace = 'backspace';
-
-  /// Identifier for the Tab key.
-  static const tab = 'tab';
-
-  /// Identifier for the Caps Lock key.
-  static const capslock = 'capslock';
-
-  /// Identifier for the Enter or Return key.
-  static const enter = 'enter';
-
-  /// Identifier for the Shift key.
-  static const shift = 'shift';
 }
