+++ packages/flutter_onscreen_keyboard/lib/src/layouts/mobile_layout.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_onscreen_keyboard/flutter_onscreen_keyboard.dart';
-import 'package:flutter_onscreen_keyboard/src/layouts/desktop_layout.dart';
+import 'package:flutter_onscreen_keyboard/src/constants/action_key_type.dart';
 import 'package:flutter_onscreen_keyboard/src/utils/extensions.dart';
 
 /// A predefined [KeyboardLayout] implementation optimized for mobile devices.
